Employee Management System

+++ employee-management-system/src/main/java/guvi/capston/employee_management_system/service/EmployeeServiceImpl.java
@@ -1,5 +1,6 @@
 package guvi.capston.employee_management_system.service;
 
+import guvi.capston.employee_management_system.ExceptionHandler.EmployeeNotFoundException;
 import guvi.capston.employee_management_system.model.Employee;
 import guvi.capston.employee_management_system.repository.EmployeeRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -31,25 +32,27 @@ public class EmployeeServiceImpl implements EmployeeService {
 	@Override
 	public Employee getEmployeeById(long id) {
 		Optional<Employee> optional = employeeRepository.findById(id);
-		Employee employee = null;
 		if (optional.isPresent()) {
-			employee = optional.get();
+			return optional.get();
 		} else {
-			throw new RuntimeException(" Employee not found for id :: " + id);
+			throw new EmployeeNotFoundException("Employee not found for id :: " + id);
 		}
-		return employee;
 	}
 
 	@Override
 	public void deleteEmployeeById(long id) {
-		this.employeeRepository.deleteById(id);
+		if (employeeRepository.existsById(id)) {
+			this.employeeRepository.deleteById(id);
+		} else {
+			throw new EmployeeNotFoundException("Employee not found for id :: " + id);
+		}
 	}
 
 	@Override
 	public Page<Employee> findPaginated(int pageNo, int pageSize, String sortField, String sortDirection) {
 		Sort sort = sortDirection.equalsIgnoreCase(Sort.Direction.ASC.name()) ? Sort.by(sortField).ascending() :
-			Sort.by(sortField).descending();
-		
+				Sort.by(sortField).descending();
+
 		Pageable pageable = PageRequest.of(pageNo - 1, pageSize, sort);
 		return this.employeeRepository.findAll(pageable);
 	}
